requirements.txt
streamlit==1.7.0
qrcode[pil]==7.3
Pillow
numpy